Remove PacmanV1
Remove PacmanV1
