+++ pacman-server/src/index.js
@@ -1,18 +1,14 @@
-import pacmanv2 from "./pacnmanv2";
-
-import pacman from "./pacman";
-
+import PacmanController from "./pacman";
 
 $(() => {
     var socket = io();
     socket.on('action', function (action) {
         let direction = action.action;
         let player = action.player
-        pacmanv2.move(player, direction);
-        pacman.move(player, direction);
+        PacmanController.move(player, direction);
     });
 });
-setTimeout(() => {
-    pacmanv2.startGameplayNoGhost();
-    pacman.startGameplayNoGhost();
-}, 2000)
+$(()=>{
+    PacmanController.startGameplayNoGhost();
+
+})
