+++ pacman-server/src/pacmanConst.js
